refactor(logs): keep one canonical formatter
DisplayTimezoneFormatter and the startup module's private formatter had
converged into the same class: both stamp in UTC and indent embedded
newlines so the collector keeps a multi-line record whole. One of them
survives, in the startup module beside the other pre-configuration
plumbing, and the display_timezone module goes. The migration-seed guard
and the UTC-stamp test move with it.
