--- dispatcharr/display_timezone.py
@@ -1,15 +1,0 @@
-import logging
-import time
-
-
-class DisplayTimezoneFormatter(logging.Formatter):
-    """Stamps records in UTC; the log collector renders the display zone."""
-
-    converter = time.gmtime
-
-    def __init__(self, format=None, datefmt=None, style="%"):
-        super().__init__(fmt=format, datefmt=datefmt, style=style)
-
-    def format(self, record):
-        # Indent embedded newlines: the collector reads leading whitespace as a continuation.
-        return super().format(record).replace("\n", "\n ")
